+++ pyaerocom/io/mscw_ctm/additional_variables.py
@@ -3,7 +3,7 @@ import logging
 import xarray as xr
 from geonum.atmosphere import T0_STD, p0
 from pyaerocom.aux_var_helpers import concx_to_vmrx
-from pyaerocom.molmasses import get_molmass
+from pyaerocom.units.molecularmass import get_molmass
 
 from pyaerocom.units.constants import M_N, M_O, M_H
