+++ pyaerocom/io/read_ebas.py
@@ -44,7 +44,7 @@ from pyaerocom.io.ebas_nasa_ames import EbasNasaAmesFile
 from pyaerocom.io.ebas_varinfo import EbasVarInfo
 from pyaerocom.io.helpers import _check_ebas_db_local_vs_remote
 from pyaerocom.io.readungriddedbase import ReadUngriddedBase
-from pyaerocom.molmasses import get_molmass
+from pyaerocom.units.molecularmass import get_molmass
 from pyaerocom.stationdata import StationData
 from pyaerocom.units.datetime import TsType
 from pyaerocom.ungriddeddata import UngriddedData
