+++ pyaerocom/units/__init__.py
@@ -1,0 +1,1 @@
+from .molecularmass import get_molmass
